+++ src/main/scala/millfork/assembly/m6809/opt/FlowAnalyzer.scala
@@ -52,10 +52,7 @@ object FlowAnalyzer {
     }
     val labelMap: () => Option[Map[String, Int]] = () => req match {
       case FlowInfoRequirement.NoRequirement => None
-      case _ => Some(code.flatMap {
-        case MLine0(op, _, MemoryAddressConstant(Label(l))) if op != MOpcode.LABEL => Some(l)
-        case _ => None
-      }.groupBy(identity).mapValues(_.size).view.force)
+      case _ => Some(code.flatMap(_.parameter.extractLabels).groupBy(identity).mapValues(_.size).view.force)
     }
     val holder = new FlowHolder(forwardFlow, reverseFlow)
     code.zipWithIndex.map{ case (line, i) => FlowInfo(holder, i, labelMap) -> line}