+++ src/test/scala/millfork/test/WarningSuite.scala
@@ -27,4 +27,18 @@ class WarningSuite extends FunSuite with Matchers {
       m.readByte(0xc000) should equal(4)
     }
   }
+  test("Loop over non-volatile variables") {
+    EmuUnoptimizedCrossPlatformRun(Cpu.Mos, Cpu.Z80, Cpu.Motorola6809)(
+      """
+        | byte flag
+        | volatile byte vflag
+        | void main () {
+        |  flag = 0
+        |  vflag = 0
+        |  while (flag != 0) {}
+        |  while (vflag != 0) {}
+        | }
+      """.stripMargin) { m =>
+    }
+  }
 }